requirements.txt
pytest>=8.0.0
pytest-playwright>=0.4.0
python-dotenv>=1.0.0
